mytravel-flask_Login/flask/requirements.txt
aniso8601==9.0.1
asgiref==3.3.4
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
Cerberus==1.3.4
cffi==1.14.5
click==8.0.0
colorama==0.4.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.17.3
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==0.16.1